@@ -33,7 +33,7 @@
   "homepage": "https://github.com/react-bootstrap/react-router-bootstrap",
   "peerDependencies": {
     "react": ">=0.13.0",
-    "react-router": ">=1.0.0-rc1"
+    "react-router": ">=1.0.0-rc3"
   },
   "devDependencies": {
     "babel": "^5.8.23",
@@ -44,21 +44,21 @@
     "colors": "^1.1.2",
     "css-loader": "^0.19.0",
     "es5-shim": "^4.1.14",
-    "eslint": "1.6.x",
-    "eslint-config-airbnb": "0.1.0",
+    "eslint": "^1.7.1",
+    "eslint-config-airbnb": "^0.1.0",
     "eslint-plugin-babel": "^2.1.1",
     "eslint-plugin-mocha": "^1.0.0",
     "eslint-plugin-react": "^3.5.1",
     "file-loader": "^0.8.4",
-    "history": "^1.12.3",
+    "history": "^1.12.5",
     "html-webpack-plugin": "^1.6.2",
-    "karma": "^0.13.10",
+    "karma": "^0.13.11",
     "karma-cli": "^0.1.1",
     "karma-mocha": "^0.2.0",
     "karma-mocha-reporter": "^1.1.1",
     "karma-phantomjs-launcher": "^0.2.1",
     "karma-sinon-chai": "^1.1.0",
-    "karma-sourcemap-loader": "^0.3.5",
+    "karma-sourcemap-loader": "^0.3.6",
     "karma-webpack": "^1.7.0",
     "less": "^2.5.3",
     "less-loader": "^2.2.1",
@@ -68,10 +68,10 @@
     "node-libs-browser": "^0.5.3",
     "phantomjs": "^1.9.18",
     "react": "^0.14.0",
-    "react-bootstrap": "^0.27.1",
+    "react-bootstrap": "^0.27.2",
     "react-dom": "^0.14.0",
     "react-router": "^1.0.0-rc3",
-    "release-script": "^0.5.3",
+    "release-script": "^0.5.4",
     "rimraf": "^2.4.3",
     "shelljs": "^0.5.3",
     "style-loader": "^0.12.4",
