@@ -19,19 +19,16 @@ describe('IndexLinkContainer', () => {
 
       describe('active state', () => {
         function renderComponent(location) {
-          class LinkWrapper extends React.Component {
-            render() {
-              return (
-                <IndexLinkContainer to="/">
-                  <Component>Root</Component>
-                </IndexLinkContainer>
-              );
-            }
-          }
-
           const router = ReactTestUtils.renderIntoDocument(
             <Router history={createMemoryHistory(location)}>
-              <Route path="/" component={LinkWrapper}>
+              <Route
+                path="/"
+                component={() => (
+                  <IndexLinkContainer to="/">
+                    <Component>Root</Component>
+                  </IndexLinkContainer>
+                )}
+              >
                 <IndexRoute />
                 <Route path="foo" />
               </Route>
